Fix passing session info & add security feature
Now database gets both the session ID and the amount of times that
session ID has completed the game and sent data to the database.
Added security feature is now both $ttime and $moves have to contain
some value other than "" to be able to put anything in the database.
This means that someone can't just go to recieve.php in their browser
and keep reloading that in order to flood the table.